Fix incorrect package object location in directory tree
